fix swiftlint errors.
